Created parallaxing background, and character classes
In the end, the Player class will be abstract and each character will
derive from it.
Worked heavily on implementing parallxing backgrounds, but I think it
looks choppy. Hard to tell with crap placeholder images(?).

I've also put in the "scale" variable that has yet to be implemented.
With zero research, I want to say that scale is taken by multiplying the
screen res height and width and dividing that by the product of "target
resolution"'s height and width.
Example: if we target 1600x900, and the user's res is 1280x720, we the
scale is .64(?)