It worked!
There was this really annoying bug where when you collide with another
object, you would kind of vibrate. I found out that I was updating
boundary, but not position, so I decided to remove the Vector2 position
variable from the player and only make references to the Rectange
boundary variable's X/Y coords instead. This fixed it. There are still
issues with collision, though.