HOLY MOTHER, I FIXED COLLISION
So, it was previously set up so that the collision boundaries (top,
bottom, left, right) would be slightly higher (+1) than the boundary of
the player. This was a start, but because the player speed is 6, I
actually had to set the collision boundaries at 7 in order for it to
work properly, otherwise, when (for example), the left and top would
collide into a corner...wait. Maybe I don't know why it works...It's
late. I've been working all day. All I know is that it works.

Still need to implement some sort of gravity, though.