Changed the structure of where Content.Load<> is called.
I gave the Level class its own LoadContent() method so that each level
can load its own content based on the first line of the map. It will
load all textures it needs to and then when it reads the tilemap, it
will use a hashtable to find which texture from the list of textures to
use.