Modified Collision for Artist Special
Previously, you couldn't jump if you were on a diagonal line of the
special because you were simultaneously colliding to the right and
bottom. This would set the canJump flag to false during the collision of
the right side. I removed that clause and made some other changes to
collision. I think we should really have a CollidableObject class.