Added "Play Map" Functionality to LE
You can now save and play the maps you make in the level editor.
To-Dos:
- It currently saves as the name "test". Implement a prompt.
- For some reason the jump functionality isn't working when you play the
map.
- Need to implement "go back to the editor" functionality when playing
the test level.
- Need to have the Buttons on the menu have text instead of creating an
image with text on it.
- Use SpriteBatch.DrawString() as part of the Button.Update() method.
- Need to make sure that clicks on the menu don't have any effect on the
player while in the testLevel.

I also started categorizing our textures. Tiles have a /Tile folder,
Buttons have a /Button Folder. This effects how the tiles are named.
Previously, it was like, "Block", now it's "Tiles/Block" so be aware of
that when checking names of textures.