Fixed Menu UI bug
Since the menu position has to be recalculated every frame regardless of
player speed due to the isHidden property, I removed boundary -= speed.
Previously this caused the menu to shift a bit in the speed direction.

I also fleshed out the Prompt class a bit.

I'm starting to think that the current Menu class needs to be specific
only to the LE menu. I think we'll have to create a custom menu class
for the pause menu and main menu as well instead of using a generic
class (probably).