Added MonoGame-specific Class Libraries
I added the packages that we'll use setting them up with MonoGame
templates. The templates are for games, but I set each output to Class
LIbrary instead of Windows Application.

I think we need to set it up so that Common has all the DLLs that we
need, like Lidgren.Network.dll, OpenTK.dll, SDL.dll and Tao.Sdl.dll.
Currently when compiling, each of these libraries is copied to every
output directory for each package. We really only need these to be
output to one package...if we set it up so that all the sub-packages
reference Common and LetsMakeAGame references Common. Does that sound
right?