Added, tested Camera Class
commented out code in Game1 and made it use LevelManager.Level,
Avatar.Players.Artist, eContentManager.eContentManager, and
Common.Camera and it worked as expected. I'll be slowly re-factoring
everything so that Game1 will be able to do things from the class
libraries instead of the classes it has in the LetsMakeAGame executable.