Fixed the skippy-ness. Introduced other issues. :/
When the player falls from a high up enough height, it will "collide"
with the left and right sensors.
I'm sure there are more issues. Can't think of them. So tired. :/
Potential fix could be to check for high velocity.Y in the wall
collision logic, but then it would clip through a wall if falling and
colliding with a wall...